Separate `mime_element` module from `mime`, some C99 cleanup
